+++ src/Cabal2Pkg/Command/Init.hs
@@ -9,7 +9,8 @@ import Cabal2Pkg.CmdLine
   ( CLI, InitOptions(..), debug, fatal, info, warn, canonPkgDir, origPkgDir
   , makeCmd, pkgBase, runMake, wantCommitMsg )
 import Cabal2Pkg.Command.Common
-  ( command, option, fetchMeta, shouldHaveBuildlink3, shouldHaveHsPrefix )
+  ( command, option, fetchMeta, shouldHaveBuildlink3, shouldHaveHsPrefix
+  , warnOutdated )
 import Cabal2Pkg.Extractor (PackageMeta(distBase))
 import Cabal2Pkg.Generator.Buildlink3 (genBuildlink3)
 import Cabal2Pkg.Generator.CommitMsg (genImportMsg)
@@ -44,6 +45,7 @@ run (InitOptions {..}) =
                , "remote file instead."
                ]
      meta <- fetchMeta pkgURI
+     warnOutdated meta
 
      canonDir <- canonPkgDir
      validatePkgPath meta