+++ src/Cabal2Pkg/Extractor/Component.hs
@@ -13,17 +13,13 @@ import Cabal2Pkg.Extractor.Conditional
   ( Environment(..), CondBlock, extractCondBlock )
 import Cabal2Pkg.Extractor.Dependency (DepSet, extractDeps)
 import Control.Monad (when, unless)
-import Data.Bifunctor (first)
 import Data.Data (Data)
 import Data.Foldable (foldl', toList)
 import Data.Map.Strict qualified as M
-import Data.Set (Set)
 import Data.Set qualified as S
 import Data.Maybe (maybeToList)
 import Data.Text (Text)
 import Data.Text qualified as T
-import Data.Traversable (mapAccumL)
-import Data.Tuple (swap)
 import Distribution.Pretty (prettyShow)
 import Distribution.Types.CondTree qualified as C
 import Distribution.Types.ConfVar qualified as C
@@ -59,35 +55,25 @@ data ComponentType
 makeLenses ''ComponentMeta
 
 
-extractComponents :: GenericPackageDescription -> CLI ([ComponentMeta], Set Text)
+extractComponents :: GenericPackageDescription -> CLI [ComponentMeta]
 extractComponents gpd
   = do env     <- extractEnv
-       lib     <- aggregateEach <$> traverse (extractLib    env      ) (GPD.condLibrary      gpd)
-       subLibs <- aggregateEach <$> traverse (extractLib    env . snd) (GPD.condSubLibraries gpd)
-       frnLibs <- aggregateEach <$> traverse (extractFrnLib env . snd) (GPD.condForeignLibs  gpd)
-       execs   <- aggregateEach <$> traverse (extractExe    env . snd) (GPD.condExecutables  gpd)
-       pure $ aggregateAll
-         [ first maybeToList lib
+       lib     <- traverse (extractLib    env      ) (GPD.condLibrary      gpd)
+       subLibs <- traverse (extractLib    env . snd) (GPD.condSubLibraries gpd)
+       frnLibs <- traverse (extractFrnLib env . snd) (GPD.condForeignLibs  gpd)
+       execs   <- traverse (extractExe    env . snd) (GPD.condExecutables  gpd)
+       pure . aggregate $
+         [ maybeToList lib
          , subLibs
          , frnLibs
          , execs
          ]
   where
-    -- Why do we keep track of Set Text manually, and don't use WriterT?
-    -- Because we need parallelisation which restricts our monad to
-    -- MonadUnliftIO.
-    aggregateAll :: [ ([ComponentMeta], Set Text) ] -> ([ComponentMeta], Set Text)
-    aggregateAll = first (filter (not . isEmpty)) . mconcat
+    aggregate :: [[ComponentMeta]] -> [ComponentMeta]
+    aggregate = filter (not . isEmpty) . mconcat
 
     isEmpty :: ComponentMeta -> Bool
     isEmpty c = c ^. cDeps . to (== mempty)
-
-    aggregateEach :: Traversable t
-                  => t (ComponentMeta, Set Text)
-                  -> (t ComponentMeta, Set Text)
-    aggregateEach = swap . mapAccumL go mempty
-      where
-        go ts0 (c, ts) = (ts0 <> ts, c)
 
     extractEnv :: CLI Environment
     extractEnv
@@ -102,68 +88,49 @@ extractComponents gpd
     pd :: C.PackageDescription
     pd = GPD.packageDescription gpd
 
-    aggregateOuter :: (CondBlock DepSet -> ComponentMeta)
-                   -> CondBlock (DepSet, Set Text)
-                   -> (ComponentMeta, Set Text)
-    aggregateOuter f = first f . swap . mapAccumL go mempty
-      where
-        go ts0 (deps, ts) = (ts0 <> ts, deps)
-
-    extractLib :: Environment -> C.CondTree C.ConfVar c C.Library -> CLI (ComponentMeta, Set Text)
+    extractLib :: Environment -> C.CondTree C.ConfVar c C.Library -> CLI ComponentMeta
     extractLib = extractCondBlock extractContent extractOuter
       where
-        extractContent :: C.Library -> Conc CLI (DepSet, Set Text)
+        extractContent :: C.Library -> Conc CLI DepSet
         extractContent = extractDeps pd . C.libBuildInfo
 
-        extractOuter :: C.Library
-                     -> CondBlock (DepSet, Set Text)
-                     -> (ComponentMeta, Set Text)
-        extractOuter lib = aggregateOuter go
-          where
-            go deps
-              = ComponentMeta
-                { _cType = Library
-                , _cName = case C.libName lib of
-                             C.LMainLibName  -> T.pack . prettyShow . C.pkgName . C.package $ pd
-                             C.LSubLibName n -> T.pack $ C.unUnqualComponentName n
-                , _cDeps = deps
-                }
+        extractOuter :: C.Library -> CondBlock DepSet -> ComponentMeta
+        extractOuter lib deps =
+          ComponentMeta
+          { _cType = Library
+          , _cName = case C.libName lib of
+                       C.LMainLibName  -> T.pack . prettyShow . C.pkgName . C.package $ pd
+                       C.LSubLibName n -> T.pack $ C.unUnqualComponentName n
+          , _cDeps = deps
+          }
 
-    extractFrnLib :: Environment -> C.CondTree C.ConfVar c C.ForeignLib -> CLI (ComponentMeta, Set Text)
+    extractFrnLib :: Environment -> C.CondTree C.ConfVar c C.ForeignLib -> CLI ComponentMeta
     extractFrnLib = extractCondBlock extractContent extractOuter
       where
-        extractContent :: C.ForeignLib -> Conc CLI (DepSet, Set Text)
+        extractContent :: C.ForeignLib -> Conc CLI DepSet
         extractContent = extractDeps pd . C.foreignLibBuildInfo
 
-        extractOuter :: C.ForeignLib
-                     -> CondBlock (DepSet, Set Text)
-                     -> (ComponentMeta, Set Text)
-        extractOuter frnLib = aggregateOuter go
-          where
-            go deps
-              = ComponentMeta
-                { _cType = ForeignLib
-                , _cName = T.pack . C.unUnqualComponentName $ C.foreignLibName frnLib
-                , _cDeps = deps
-                }
+        extractOuter :: C.ForeignLib -> CondBlock DepSet -> ComponentMeta
+        extractOuter frnLib deps =
+          ComponentMeta
+          { _cType = ForeignLib
+          , _cName = T.pack . C.unUnqualComponentName $ C.foreignLibName frnLib
+          , _cDeps = deps
+          }
 
-    extractExe :: Environment -> C.CondTree C.ConfVar c C.Executable -> CLI (ComponentMeta, Set Text)
+    extractExe :: Environment -> C.CondTree C.ConfVar c C.Executable -> CLI ComponentMeta
     extractExe = extractCondBlock extractContent extractOuter
       where
-        extractContent :: C.Executable -> Conc CLI (DepSet, Set Text)
+        extractContent :: C.Executable -> Conc CLI DepSet
         extractContent = extractDeps pd . C.buildInfo
 
-        extractOuter :: C.Executable
-                     -> CondBlock (DepSet, Set Text)
-                     -> (ComponentMeta, Set Text)
-        extractOuter exe = aggregateOuter go
-          where
-            go deps
-              = ComponentMeta
-                { _cType = Executable
-                , _cName = T.pack . C.unUnqualComponentName $ C.exeName exe
-                , _cDeps = deps
-                }
+        extractOuter :: C.Executable -> CondBlock DepSet -> ComponentMeta
+        extractOuter exe deps =
+          ComponentMeta
+          { _cType = Executable
+          , _cName = T.pack . C.unUnqualComponentName $ C.exeName exe
+          , _cDeps = deps
+          }
 
 
 extractFlags :: FlagMap -> [C.PackageFlag] -> CLI FlagMap