+++ src/Cabal2Pkg/Extractor/Dependency.hs
@@ -11,9 +11,6 @@ import Cabal2Pkg.Extractor.Dependency.ExternalLib (ExtLibDep, extractExtLibDep)
 import Cabal2Pkg.Extractor.Dependency.Library (LibDep, extractLibDep)
 import Cabal2Pkg.Extractor.Dependency.PkgConfig (PkgConfDep, extractPkgConfDep)
 import Data.Data (Data)
-import Data.Set (Set)
-import Data.Set qualified as S
-import Data.Text (Text)
 import Distribution.Simple.BuildToolDepends qualified as BTD
 import Distribution.Types.BuildInfo qualified as C
 import Distribution.Types.Dependency qualified as C
@@ -43,29 +40,12 @@ data DepSet
 makeLenses ''DepSet
 
 
--- |The second element of the tuple is a set of Cabal packages that need to
--- be listed in @HASKELL_UNRESTRICT_DEPENDENCIES@.
-extractDeps :: C.PackageDescription -> C.BuildInfo -> Conc CLI (DepSet, Set Text)
+extractDeps :: C.PackageDescription -> C.BuildInfo -> Conc CLI DepSet
 extractDeps pkg bi
-  = aggregateAll <$> execs <*> extLibs <*> libs <*> pkgConfLibs
+  = DepSet <$> execs <*> extLibs <*> libs <*> pkgConfLibs
   where
-    aggregateAll :: ([ExeDep], Set Text)
-                 -> [ExtLibDep]
-                 -> ([LibDep], Set Text)
-                 -> [PkgConfDep]
-                 -> (DepSet, Set Text)
-    aggregateAll (eds, ts0) elds (lds, ts1) pcds
-      = (DepSet eds elds lds pcds, ts0 <> ts1)
-
-    aggregateEach :: [(Maybe a, Maybe Text)] -> ([a], Set Text)
-    aggregateEach = foldr go ([], mempty)
-      where
-        go (a, b) (as, bs) = ( maybe as (: as) a
-                             , maybe bs (`S.insert` bs) b
-                             )
-
-    execs :: Conc CLI ([ExeDep], Set Text)
-    execs = aggregateEach <$> traverse (conc . extractExeDep)
+    execs :: Conc CLI [ExeDep]
+    execs = traverse (conc . extractExeDep)
             [ dep
             | dep <- BTD.getAllToolDependencies pkg bi
             , not (BTD.isInternal pkg dep)
@@ -74,8 +54,8 @@ extractDeps pkg bi
     extLibs :: Conc CLI [ExtLibDep]
     extLibs = pure . (extractExtLibDep <$>) $ C.extraLibs bi
 
-    libs :: Conc CLI ([LibDep], Set Text)
-    libs = aggregateEach <$> traverse (conc . extractLibDep)
+    libs :: Conc CLI [LibDep]
+    libs = traverse (conc . extractLibDep)
            [ dep
            | dep <- C.targetBuildDepends bi
            , not (isInternalLib dep)