+++ src/Cabal2Pkg/Extractor/Dependency/Executable.hs
@@ -6,7 +6,6 @@ module Cabal2Pkg.Extractor.Dependency.Executable
   ) where
 
 import Cabal2Pkg.CmdLine (CLI, progDb, srcDb)
-import Control.Monad (join)
 import Data.Data (Data)
 import Data.Text (Text)
 import Data.Text qualified as T
@@ -16,63 +15,77 @@ import Distribution.Parsec (eitherParsec)
 import Distribution.Simple.Program.Db qualified as C
 import Distribution.Simple.Program.Types qualified as C
 import Distribution.Types.ExeDependency qualified as C
+import Distribution.Types.PackageName (PackageName)
 import Distribution.Types.PackageName qualified as C
 import Distribution.Types.Version (Version)
-import Distribution.Types.VersionRange qualified as C
+import Distribution.Types.VersionRange (VersionRange)
 import System.OsPath.Posix (pstr)
 import System.OsPath.Posix qualified as OP
 
 
 -- |Dependency on a tool provided by a pkgsrc package.
 data ExeDep
-  = KnownExe
-    { -- |The name of a tool to be listed in @USE_TOOLS@.
-      name :: !Text
+  = KnownBundledExe
+    { -- |The name of tool, such as @"hsc2hs"@. This constructor is used
+      -- when 'Cabal2Pkg.Extractor.summariseCabal' finds that the
+      -- dependency is bundled with the compiler.
+      name :: !PackageName
+      -- |The version of the tool it found.
+    , version :: !Version
+      -- |The range of versions that is acceptable.
+    , verRange :: !VersionRange
+    }
+  | KnownPkgsrcExe
+    { -- |The name of a tool to be listed in @USE_TOOLS@. This constructor
+      -- is used when 'Cabal2Pkg.Extractor.summariseCabal' finds that the
+      -- dependency is packaged in pkgsrc.
+      name :: !PackageName
+      -- |The PKGPATH, such as @"devel/happy"@.
+    , pkgPath :: !Text
+      -- |The version of the tool it found.
+    , version :: !Version
+      -- |The range of versions that is acceptable.
+    , verRange :: !VersionRange
     }
   | UnknownExe
     { -- |The name of Cabal package, such as @"alex"@. This constructor is
       -- used when 'Cabal2Pkg.Extractor.summariseCabal' cannot find the
       -- corresponding package in pkgsrc.
-      name :: !Text
+      name :: !PackageName
     }
   deriving (Data, Eq, Show)
 
 
--- |Return @(md, mt)@ where @md@ is 'Nothing' if the dependency is bundled
--- with the compiler, and @mt@ is the name of Cabal package if it needs to
--- be listed in @HASKELL_UNRESTRICT_DEPENDENCIES@.
-extractExeDep :: C.ExeDependency -> CLI (Maybe ExeDep, Maybe Text)
-extractExeDep (C.ExeDependency pkgName _ verRange)
+extractExeDep :: C.ExeDependency -> CLI ExeDep
+extractExeDep (C.ExeDependency pkgName _ range)
   = do progs <- progDb
        case lookupBundled progs pkgName of
          Just ver -> pure $ bundled ver
          Nothing  ->
            do m <- findPkgsrcPkg pkgName
               case m of
-                Just (name, ver) -> pure $ found name ver
+                Just (path, ver) -> pure $ found path ver
                 Nothing          -> pure notFound
   where
-    name' :: Text
-    name' = T.pack . C.unPackageName $ pkgName
+    bundled :: Version -> ExeDep
+    bundled ver =
+      KnownBundledExe
+      { name     = pkgName
+      , version  = ver
+      , verRange = range
+      }
 
-    needsU :: Version -> Maybe Text
-    needsU ver
-      | not $ C.withinRange ver verRange
-          = Just name'
-      | otherwise
-          = Nothing
+    found :: Text -> Version -> ExeDep
+    found path ver =
+      KnownPkgsrcExe
+      { name     = pkgName
+      , pkgPath  = path
+      , version  = ver
+      , verRange = range
+      }
 
-    bundled :: Version -> (Maybe ExeDep, Maybe Text)
-    bundled ver
-      = (Nothing, needsU ver)
-
-    found :: Text -> Version -> (Maybe ExeDep, Maybe Text)
-    found name ver
-      = (Just (KnownExe name), needsU ver)
-
-    notFound :: (Maybe ExeDep, Maybe Text)
-    notFound
-      = (Just (UnknownExe name'), Nothing)
+    notFound :: ExeDep
+    notFound = UnknownExe { name = pkgName }
 
 lookupBundled :: C.ProgramDb -> C.PackageName -> Maybe Version
 lookupBundled progs pkgName
@@ -83,8 +96,7 @@ lookupBundled progs pkgName
 -- |Search for a pkgsrc package case-insensitively, both with and without
 -- the @hs-@ prefix. Whether it includes @mk/haskell.mk@ or not is
 -- irrelevant because we don't care in what language it's implemented. The
--- function returns a pair of the name of the tool to be listed in
--- @USE_TOOLS@ and its @PKGVERSION_NOREV@.
+-- function returns a pair of its @PKGPATH@ and @PKGVERSION_NOREV@.
 findPkgsrcPkg :: C.PackageName -> CLI (Maybe (Text, Version))
 findPkgsrcPkg name
   = do db <- srcDb
@@ -95,15 +107,16 @@ findPkgsrcPkg name
        name' <- OP.encodeUtf . C.unPackageName $ name
        p0    <- SrcDb.findPackageByNameCI db name'
        case p0 of
-         Just p  -> found p
+         Just p  -> Just <$> found p
          Nothing ->
            do p1 <- SrcDb.findPackageByNameCI db ([pstr|hs-|] <> name')
-              join <$> traverse found p1
+              traverse found p1
   where
-    found :: Package CLI -> CLI (Maybe (Text, Version))
-    found pkg
-      = do ver <- toCabalVer =<< SrcDb.pkgVersionNoRev pkg
-           pure $ Just (T.pack . C.unPackageName $ name, ver)
+    found :: Package CLI -> CLI (Text, Version)
+    found pkg =
+      do ver  <- toCabalVer =<< SrcDb.pkgVersionNoRev pkg
+         path <- T.pack <$> OP.decodeUtf (SrcDb.pkgPath pkg)
+         pure (path, ver)
 
     toCabalVer :: MonadFail m => Text -> m Version
     toCabalVer = either fail pure . eitherParsec . T.unpack